day 6: drawSlime api cleanup, sweat timing
Sweat removed after releasing jump button. Add more sweat. Make sweat
invisible for the first couple of ms and put sweat in foreground instead
of behind slime. Add level drawing mechanic, but nothing to show for it
yet.

Looking to add a proper indicator graphic for your targetting system,
instead of the hideous bezier curve that is currently in the system.